pyproject.toml
[project]
name = "rag-system-scientific-articles"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "docling>=2.84.0",
    "google-genai>=1.70.0",
    "pymupdf>=1.27.2.2",
    "pyyaml>=6.0.3",
]

[project.scripts]
rag = "main:main"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]

[tool.uv]
package = true

[tool.hatch.build.targets.wheel]
packages = ["src"]
